requirements.txt
flask
crewai
langchain-groq
crewai-tools
python-dotenv
langchain-huggingface
google-generativeai
flask-cors
